Public inboxes are writable by other instances
This may be an unnecessary tightening of the spec. It may be a minor information leak if other instances can _read_ from the public inbox. 

(Note: I am assuming we want to standardize on using the word "Instance" over "server" in the protocol specification except in cases where we literally mean a single virtual or physical server. I am happy to include that separately if you prefer)